+++ colors.js
@@ -1,0 +1,17 @@
+const MajorColors = [
+    "white",
+    "red",
+    "black",
+    "yellow",
+    "violet"
+];
+
+const MinorColors = [
+    "blue",
+    "orange",
+    "green",
+    "brown",
+    "slate"
+];
+
+module.exports={MajorColors,MinorColors};
