+++ utils.js
@@ -1,0 +1,45 @@
+const expect = require('chai').expect;
+const colors=require('./colors');
+const MajorColors = colors.MajorColors;
+const MinorColors = colors.MinorColors;
+
+function GetColorFromPairNumber(pairNumber) {
+    let colorPair = {};
+    const zeroBasedPairNumber = pairNumber - 1;
+    const majorColorIndex =
+        Math.floor(zeroBasedPairNumber / MinorColors.length);
+    colorPair.major = MajorColors[majorColorIndex];
+    colorPair.minor = MinorColors[zeroBasedPairNumber % MinorColors.length];
+    return colorPair
+}
+
+function GetPairNumberFromColors(majorColor, minorColor) {
+    let majorIndex = 0;
+    let minorIndex = 0;
+    for(majorIndex = 0; majorIndex < MajorColors.length; majorIndex++) {
+        if(MajorColors[majorIndex] == majorColor) {
+            break;
+        }
+    }
+    for(minorIndex = 0; minorIndex < MinorColors.length; minorIndex++) {
+        if(MinorColors[minorIndex] == minorColor) {
+            break;
+        }
+    }
+    return majorIndex * MinorColors.length + minorIndex + 1;
+}
+
+function testNumberToPair(number, expectedMajor, expectedMinor) {
+    const pairOfColors = GetColorFromPairNumber(number);
+    console.log(`${number} = ${expectedMajor} ${expectedMinor}`);
+    expect(pairOfColors.major).equals(expectedMajor);
+    expect(pairOfColors.minor).equals(expectedMinor);
+}
+
+function testColorToNumber(majorColor, minorColor, expectedNumber) {
+    const pairNumber = GetPairNumberFromColors(majorColor, minorColor);
+    console.log(`${majorColor} ${minorColor} = ${pairNumber}`);
+    expect(pairNumber).to.equal(expectedNumber);
+}
+
+module.exports={GetColorFromPairNumber,GetPairNumberFromColors,testNumberToPair,testColorToNumber};
